Crate remove_dir_all

Reliably remove a directory and all of its children.

This library provides a reliable implementation of remove_dir_all for Windows. For Unix systems, it re-exports std::fs::remove_dir_all.

It also provides remove_dir_contents and ensure_empty_dir for both Unix and Windows.

Functions

ensure_empty_dir

Makes path an empty directory: if it does not exist, it is created it as an empty directory (as if with std::fs::create_dir); if it does exist, its contents are deleted (as if with remove_dir_contents).

remove_dir_all

Removes a directory at this path, after removing all its contents. Use carefully!

remove_dir_contents

Deletes the contents of path, but not the directory iteself.