Function remove_dir_all::remove_dir_all

pub fn remove_dir_all<P>(path: P) -> Result<(), Error>where
    P: AsRef<Path>,

Removes a directory at this path, after removing all its contents. Use carefully!

This function does not follow symbolic links and it will simply remove the symbolic link itself.

Platform-specific behavior

This function currently corresponds to openat, fdopendir, unlinkat and lstat functions on Unix (except for macOS before version 10.10 and REDOX) and the CreateFileW, GetFileInformationByHandleEx, SetFileInformationByHandle, and NtCreateFile functions on Windows. Note that, this may change in the future.

On macOS before version 10.10 and REDOX, as well as when running in Miri for any target, this function is not protected against time-of-check to time-of-use (TOCTOU) race conditions, and should not be used in security-sensitive code on those platforms. All other platforms are protected.

Errors

See fs::remove_file and fs::remove_dir.

Examples

use std::fs;

fn main() -> std::io::Result<()> {
    fs::remove_dir_all("/some/dir")?;
    Ok(())
}