Crate remove_dir_all

Reliably remove a directory and all of its children.

This library provides a reliable implementation of remove_dir_all for Windows. For Unix systems, it re-exports std::fs::remove_dir_all.

Functions

remove_dir_all

Removes a directory at this path, after removing all its contents. Use carefully!