Function remove_dir_all::remove_dir_all

pub fn remove_dir_all<P>(path: P) -> Result<(), Error> where
    P: AsRef<Path>, 

Removes a directory at this path, after removing all its contents. Use carefully!

This function does not follow symbolic links and it will simply remove the symbolic link itself.

Platform-specific behavior

This function currently corresponds to opendir, lstat, rm and rmdir functions on Unix and the FindFirstFile, GetFileAttributesEx, DeleteFile, and RemoveDirectory functions on Windows. Note that, this may change in the future.

Errors

See file::remove_file and fs::remove_dir.

Examples

use std::fs;

fn main() -> std::io::Result<()> {
    fs::remove_dir_all("/some/dir")?;
    Ok(())
}